select analyzer
